requirements.txt
# hormuz-ph-urea: Strait of Hormuz closure scenario simulation
# Philippine urea supply and price impact model

# Data fetching
requests>=2.28.0
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
comtradeapicall>=1.3.0

# Data processing and modeling
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
